corregida la conversion a c++ para todos los casos de prueba de psexport

+++ test/export/cpp/06-segun.cpp
@@ -1,15 +1,9 @@
-// Este codigo ha sido generado por el modulo psexport 20131129-l64 de PSeInt
-// dado que dicho modulo se encuentra aun en desarrollo y en etapa experimental
-// puede que el codigo generado no sea completamente correcto. Si encuentra
-// errores por favor reportelos en el foro (http://pseint.sourceforge.net).
-
 #include<iostream>
 using namespace std;
-
 int main() {
-	float c;
+	int c;
 	cin>>c;
-	switch (int(c)) {
+	switch (c) {
 	case 1:
 		cout<<"1"<<endl;
 		break;
@@ -25,4 +19,3 @@ int main() {
 	}
 	return 0;
 }
-
